added java stream api operators.
How many male and female employees are there in the organization?  : done 
Print the name of all departments in the organization? :done 
What is the average age of male and female employees? : Done 
Get the details of highest paid employee in the organization? Done
Get the names of all employees who have joined after 2015? : Done
Count the number of employees in each department? : Done
What is the average salary of each department? : Done
Get the details of youngest male employee in the product development department? Done
Who has the most working experience in the organization?: Done
How many male and female employees are there in the sales and marketing team? : Done
What is the average salary of male and female employees? Done
List down the names of all employees in each department? Done
What is the average salary and total salary of the whole organization? Separate the employees who are younger or equal to 25 years from those
employees who are older than 25 years? done
Who is the oldest employee in the organization? What is his age and which department he belongs to?
How to find only duplicate elements with its count from the String ArrayList in Java8?